change button palcement
