+++ frontend/src/components/modules/Profile/EditInfoButtonRow.tsx
@@ -1,9 +1,15 @@
-import React from "react";
+import React, { useState } from "react";
 import type { UseFormReturn } from "react-hook-form";
 import type { EditUserRequest } from "types/requests";
 import Button from "@mui/material/Button";
 import Box from "@mui/material/Box";
 import { useTranslation } from "react-i18next";
+import ConfirmUserDeletionModal from "./ConfirmUserDeleteModal";
+import routePaths from "routes/route-paths";
+import useToast from "hooks/useToast";
+import { useAuth } from "context/AuthContext";
+import api from "api/axios";
+import { useNavigate } from "react-router-dom";
 
 interface EditButtonRowProps {
   editingEnabled: boolean;
@@ -17,6 +23,26 @@ const EditButtonRow: React.FC<EditButtonRowProps> = ({
   formContext
 }: EditButtonRowProps) => {
   const { t } = useTranslation();
+  const showToast = useToast();
+  const navigate = useNavigate();
+  const [isConfirmationDialogOpen, setConfirmationDialogOpen] = useState(false);
+  const { userId, logout } = useAuth();
+
+  const handleDeleteUser = async (): Promise<void> => {
+    try {
+      // This whole component wont be rendered if this is undefined.
+      // eslint-disable-next-line  @typescript-eslint/no-non-null-assertion
+      await api.user.delete(userId!);
+      await logout();
+      showToast(t("messages.deletion_success"), "success");
+      navigate(routePaths.homeRoute, {
+        replace: true,
+        state: { refresh: true }
+      });
+    } catch (error) {
+      showToast(error, "error");
+    }
+  };
 
   return (
     <Box
@@ -25,34 +51,14 @@ const EditButtonRow: React.FC<EditButtonRowProps> = ({
       flexWrap="wrap"
       gap="10px"
     >
-      {!editingEnabled ? (
-        <Button
-          type="button"
-          variant="outlined"
-          color="primary"
-          sx={{ mt: 3, mb: 2 }}
-          onClick={() => {
-            setEditingEnabled(true);
-          }}
-        >
-          {t("buttons.edit_info_button")}
-        </Button>
-      ) : (
-        <Button
-          type="button"
-          variant="outlined"
-          color="primary"
-          onClick={() => {
-            setEditingEnabled(() => {
-              formContext.reset();
-              return false;
-            });
-          }}
-          sx={{ mt: 3, mb: 2 }}
-        >
-          {t("buttons.cancel_button")}
-        </Button>
-      )}
+      <ConfirmUserDeletionModal
+        isOpen={isConfirmationDialogOpen}
+        onClose={() => {
+          setConfirmationDialogOpen(false);
+        }}
+        onConfirm={handleDeleteUser}
+      />
+
       <Button
         type="submit"
         variant="contained"
@@ -61,6 +67,18 @@ const EditButtonRow: React.FC<EditButtonRowProps> = ({
         sx={{ mt: 3, mb: 2 }}
       >
         {t("buttons.save_info_button")}
+      </Button>
+      <Button
+        type="button"
+        variant="contained"
+        color="error"
+        size="small"
+        sx={{ mt: 3, mb: 2 }}
+        onClick={() => {
+          setConfirmationDialogOpen(true);
+        }}
+      >
+        {t("buttons.delete_account_button")}
       </Button>
     </Box>
   );
